+++ trunk/src/protocol/srs_protocol_stream.cpp
@@ -167,7 +167,6 @@ srs_error_t SrsFastStream::grow(ISrsReader* reader, int required_size)
          * to improve read performance, merge some packets then read,
          * when it on and read small bytes, we sleep to wait more data.,
          * that is, we merge some data to read together.
-         * @see https://github.com/ossrs/srs/issues/241
          */
         if (merged_read && _handler) {
             _handler->on_read(nread);